bot/requirements.txt
aiogram>=3.0.0b7
sqlalchemy>=1.4
asyncpg
aiohttp
python-dotenv
greenlet>=2.0.0
uvicorn
fastapi
tronpy>=0.5.0
httpx
validators